Implement repository cloning and automatic test running (#24)
Add code that runs when webhook is triggered.
The code automatically clones the repository and builds the project when a 
pull request is made to a specific branch. It will also run the JUnit tests automatically.

Right now the program just prints the result of the compilation and tests to stdout, 
this information should be saved to a file. Will create an issue for this.

Working #22 #4