Add website frontend for displaying ci results (#27)
Implement a website frontend for the CI server, using Vue and Vuetify.
This allows for a user friendly experience to view the CI server results.
Also readied for implementation of CI server history. This feature
counts as our "something remarkable" in P+.

Next step is to save files on the server so they can be requested
to view CI server history #25. 

Working #12, #11