Add Notification system via GitHub API (#29)
This will link the CI server to GitHub in realtime so it can display if the commit 
was successful and the server can compile & run the tests.

There are some features left to fix, such as that the server can detect when 
a test fails. Also, add a link to the website via the ´details´ hyperlink. Will create 
issues for these later.

Working #6 

---------

Co-authored-by: Elliot <edarth@kth.se>