Fix compile boolean result bug (#39)
The bug was caused by how the compileState variable was set. Since there is one try-block, we can sometimes receive duplicate exceptions that will set compileState to false, even if it should not have been set.

The solution was to create a new ComplieException class that handles all issues with the compile function. This ensures that compileState is only modified when the compile fails, and not when any other error occurs in another function!

Closes #34