add secret folder to repository
add the secret folder to repository, but ignore the content of the folder with additional .gitignore file.

this is to make it easier for future developers to add their git-token to the correct folder, instead of having to figure out where to put this secret folder through trail and error.

Closes #33